Principal - limpieza de código obsoleto y comentarios. DetallePersonaje - interfaz rediseñada para optimiar el espacio. Añadida la funcionalidad para ver el detalle del hechizo seleccionado. Cambiada de lugar (de Principal a DetallePersonaje) la funcionalidad necesaria para generar los hechizos conocidos. Hechizo - cambios menores en nombres de atributos y métodos.

+++ LibroHechizos/src/principal/BolaDeFuego.java
@@ -2,7 +2,7 @@ package principal;
 
 public class BolaDeFuego extends Hechizo implements Hiriente, Mostrable{
 
-	public BolaDeFuego(short nivelLanzador) {
+	public BolaDeFuego(int nivelLanzador) {
 		super(
 				"Bola de Fuego",
 				Escuela.EVOCACION,
